Minor fix
Canceling an invalid event is part of an older commit. It was meant to be useful for blocking the "invalid" event triggered by a required, empty control.
This has already been solved by altering the "required" IDL attribute, but I saved this feature in case ReqAlt is used together with a validation polyfill which does not rely upon a (custom) "required" property. Now it is fixed accordingly.